Added a test
